Refactored attributes tests
