+++ ngraph/test/visitors/op/lrn.cpp
@@ -1,0 +1,40 @@
+// Copyright (C) 2018-2021 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#include "gtest/gtest.h"
+
+#include "ngraph/ngraph.hpp"
+#include "ngraph/op/util/attr_types.hpp"
+#include "ngraph/opsets/opset1.hpp"
+#include "ngraph/opsets/opset3.hpp"
+#include "ngraph/opsets/opset4.hpp"
+#include "ngraph/opsets/opset5.hpp"
+
+#include "util/visitor.hpp"
+
+using namespace std;
+using namespace ngraph;
+using ngraph::test::NodeBuilder;
+using ngraph::test::ValueMap;
+
+TEST(attributes, lrn_op)
+{
+    NodeBuilder::get_ops().register_factory<opset1::LRN>();
+    const auto arg = make_shared<op::Parameter>(element::f32, Shape{1, 2, 3, 4});
+    const auto axes = make_shared<op::Parameter>(element::i32, Shape{2});
+
+    const double alpha = 1.1;
+    const double beta = 2.2;
+    const double bias = 3.3;
+    const size_t size = 4;
+
+    const auto lrn = make_shared<opset1::LRN>(arg, axes, alpha, beta, bias, size);
+    NodeBuilder builder(lrn);
+    auto g_lrn = as_type_ptr<opset1::LRN>(builder.create());
+
+    EXPECT_EQ(g_lrn->get_alpha(), lrn->get_alpha());
+    EXPECT_EQ(g_lrn->get_beta(), lrn->get_beta());
+    EXPECT_EQ(g_lrn->get_bias(), lrn->get_bias());
+    EXPECT_EQ(g_lrn->get_nsize(), lrn->get_nsize());
+}
